refactor: 공통 함수, 타입, 상수 분리
