+++ types/ImageType.ts
@@ -1,0 +1,6 @@
+export interface ImageType {
+    uri: string;
+    fileName: string;
+    fileSize: number;
+    uploadDate: Date;
+}
